Update to playGame function
I love this! Such a great use of arrow functions and functional programming. 

I noticed you added an if-else statement in your `getUserFunction`! One thing I recently learned about functions that don't have a return statement is that they actually return `undefined`. So in the event someone passes in "foo", the function will return `undefined` but the rest of your `playGame` function will continue. 

By adding this if-statement, the `playGame` function will stop executing the rest of it's code.